refactoring timer to small functions
